DSL_CHAT_BOT/backend/requirements.txt
torch==2.6.0
torchvision==0.21.0  
torchaudio==2.6.0
transformers==4.45.0
sentence-transformers==3.0.1
fastapi
uvicorn
pandas
numpy
requests
python-dotenv
accelerate